pyproject.toml
[project]
name = "mcp-misp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.13.1",
    "pymisp>=2.5.17",
    "python-dotenv>=1.1.1",
]
